+++ CHANGES.rst
@@ -13,6 +13,8 @@ Unreleased
 -   Use modern packaging metadata with ``pyproject.toml`` instead of ``setup.cfg``.
     :pr:`1793`
 -   Use ``flit_core`` instead of ``setuptools`` as build backend.
+-   Fix ``indent`` filter not respecting ``blank=False`` for the first line
+    when ``first=True`` is set. :issue:`2176`
 
 
 Version 3.1.6